examples/requirements.txt
pyserial==3.4
colorama==0.4.3
